examples/chatssm-vinh/pyproject.toml
[tool.poetry]
name = "chatssm"
version = "0.0.1"
description = "ChatSSM Sandbox"
authors = ["ctn <ctn@aitomatic.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.scripts]
app = 'app:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
openai = "^0.27.0"
flask = "^2.0.0"
python-dotenv = "^0.19.0"
llama-index = ">=0.6.9"
llama-hub = ">=0.0.1"
